+++ routes/TypePaymentRecord.php
@@ -17,8 +17,8 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth', 'verified'])->get('/type_payment_record', [TypePaymentRecordController::class, 'index'])->name('type_payment_record');
 
 Route::prefix('/type_payment_record')->middleware(['auth', 'verified'])->group(function () {
-    Route::put('/', [TypePaymentRecordController::class, 'store']);
-    Route::delete('/{id}/{page}', [TypePaymentRecordController::class, 'destroy']);
+    Route::put('/', [TypePaymentRecordController::class, 'store'])->middleware('permission:type_payment_record.registrar|type_payment_record.editar');
+    Route::delete('/{id}/{page}', [TypePaymentRecordController::class, 'destroy'])->middleware('permission:type_payment_record.eliminar');
     Route::get('/tables', [TypePaymentRecordController::class, 'tables']);
     Route::get('/record/{id}', [TypePaymentRecordController::class, 'record']);
 });
