+++ routes/cities.php
@@ -19,8 +19,8 @@ $controlador = CityController::class;
 Route::middleware(['auth', 'verified'])->get('/cities', [$controlador, 'index'])->name('cities');
 
 Route::prefix('/cities')->middleware(['auth', 'verified'])->group(function () use ($controlador) {
-    Route::put('/', [$controlador, 'store']);
-    Route::delete('/{id}/{page}', [$controlador, 'destroy']);
+    Route::put('/', [$controlador, 'store'])->middleware('permission:cities.registrar|cities.editar');
+    Route::delete('/{id}/{page}', [$controlador, 'destroy'])->middleware('permission:cities.eliminar');
     Route::get('/tables', [$controlador, 'tables']);
     Route::get('/record/{id}', [$controlador, 'record']);
 });
